support multiple calendars and specify timezone

+++ sync.py
@@ -1,10 +1,13 @@
 import os
 import sys
 import time
+import re
+from typing import Set
 
 import arrow
 import caldav
 import ics
+from ics.icalendar import Event
 import requests
 
 
@@ -51,7 +54,7 @@ class ICSToCalDAV:
             ).text
         )
 
-    def _get_local_events_ids(self) -> set[int]:
+    def _get_local_events_ids(self) -> Set[int]:
         """
         This piece of crap:
         1) Gets from the local calendar all the events ocurring after now,
@@ -79,30 +82,89 @@ END:VCALENDAR
 """
 
     def synchronise(self):
+        print('.synchronise')
         """
         The main function which:
         1) Pulls all the events from the remote calendar,
         2) Saves them into the local calendar,
         3) Removes local events which are not in the remote any more.
         """
-        for remote_event in self.remote_calendar.events:
-            if arrow.utcnow() > remote_event.end:
-                continue
+        
+        print(f'found total {len(self.remote_calendar.events)} events')
+        current_time = arrow.utcnow()
+        future_events = []
+        past_events = []
 
-            self.local_calendar.save_event(self._wrap(remote_event))
-            print("+", end="")
-            sys.stdout.flush()
+        # https://icspy.readthedocs.io/en/stable/api.html#event
+        for event in self.remote_calendar.events:
+            if current_time > event.end:
+                past_events.append(event)
+            else:
+                future_events.append(event)
+
+        # sort
+        future_events = sorted(future_events, key=lambda event: event.begin)
+        past_events = sorted(past_events, key=lambda event: event.begin)
+        
+        print(f'past:{len(past_events)}, future:{len(future_events)}')
+
+        for remote_event in future_events:
+            # print()
+            # print()
+
+            # print(remote_event)
+
+            # print()
+            # print()
+            try:
+                if arrow.utcnow() > remote_event.end:
+                    continue
+                
+                print(f".processing event: {remote_event.name} ({remote_event.begin})")
+                event_str = str(remote_event)
+                new_uid = remote_event.uid + '-' + str(remote_event.begin.timestamp)
+                event_str = re.sub(
+                    r"UID:[^\n]+",  # Matches "UID:" followed by any characters except newline
+                    f"UID:{new_uid}",  # The replacement string
+                    event_str  # The string to search
+                )
+                event_str = re.sub(r"RECURRENCE-ID.*\n", "", event_str) # replace recurrence line
+
+                success = False
+                attempts = 0
+                while not success and attempts < 5:
+                    try:
+                        self.local_calendar.save_event(self._wrap(event_str))
+                        print("+", end="")
+                        sys.stdout.flush()
+                        success = True
+                    except Exception as e:
+                        attempts += 1
+                        print(f"Failed to save event {remote_event.uid} (attempt {attempts}): {e}")
+                        time.sleep(5)  # wait for 5 seconds before trying again
+            except Exception as e:
+                print(f"Failed to process event {remote_event.uid}: {e}")
+                continue
         print()
 
         remote_events_ids = set(e.uid for e in self.remote_calendar.events)
-        events_to_delete = self._get_local_events_ids() - remote_events_ids
-        for local_event_id in events_to_delete:
-            self.local_client.delete(
-                f"{self.local_calendar.url}{local_event_id}.ics"
-            )
-            print("-", end="")
-            sys.stdout.flush()
+        # events_to_delete = self._get_local_events_ids() - remote_events_ids
+        for local_event_id in self._get_local_events_ids():
+            original_uid = local_event_id.split('-')[0]  # assuming the uid format is uid-starttime
+            if original_uid not in remote_events_ids:
+                self.local_client.delete(
+                    f"{self.local_calendar.url}{local_event_id}.ics"
+                )
+                print("-", end="")
+                sys.stdout.flush()
         print()
+
+        # for local_event_id in events_to_delete:
+        #     self.local_client.delete(
+        #         f"{self.local_calendar.url}{local_event_id}.ics"
+        #     )
+        #     print("-", end="")
+        #     sys.stdout.flush()
 
 
 def getenv_or_raise(var):
@@ -112,6 +174,7 @@ def getenv_or_raise(var):
 
 
 if __name__ == "__main__":
+    print('.set up env')
     settings = {
         "remote_url": getenv_or_raise("REMOTE_URL"),
         "local_url": getenv_or_raise("LOCAL_URL"),
@@ -125,6 +188,7 @@ if __name__ == "__main__":
     sync_every = os.getenv("SYNC_EVERY", None)
     if sync_every is not None:
         sync_every = "in " + sync_every
+        print(f'.sync every{sync_every}')
         try:
             arrow.utcnow().dehumanize(sync_every)
         except ValueError as ve:
